# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soccer Api Documentation",
        "description": "Soccer API for doing CURD operation.",
        "contact": {
            "email": "b.sahoo@devon.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost.soccer.com/api",
            "description": "Soccer API Server"
        },
        {
            "url": "localhost.soccer.com/api"
        }
    ],
    "paths": {
...
        "/player/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get a player details",
                "description": "Returns list of players",
                "operationId": "getPlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Delete existing player",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Player deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "player/{id}": {
            "put": {
                "tags": [
                    "Player"
                ],
                "summary": "Update existing player",
                "description": "Update player data",
                "operationId": "updatePlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Players updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/teams": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Get list of teams",
                "description": "Returns list of teams",
                "operationId": "getTeamList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/team/{id}": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Get team details",
                "description": "Returns details of the team",
                "operationId": "getTeam",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Team"
                ],
                "summary": "Delete existing team",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/team": {
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Store new Team",
                "description": "To cerate a new team",
                "operationId": "storeTeam",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTeamRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "team/{id}": {
            "put": {
                "tags": [
                    "Team"
                ],
                "summary": "Update existing Team",
                "description": "Update Team data",
                "operationId": "updateTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTeamRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Players updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Player": {
                "title": "Player",
                "description": "Player model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name of the player",
                        "type": "string",
                        "example": "Lionel"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name of the player",
                        "type": "string",
                        "example": "Messi"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image file path of the player",
                        "type": "string",
                        "example": "/playre123625.jpg"
                    },
                    "team_id": {
                        "title": "Team Id",
                        "description": "Player's team ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "3"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Player"
                }
            },
            "Team": {
                "title": "Team",
                "description": "Team model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the Team",
                        "type": "string",
                        "example": "unity"
                    },
                    "logo": {
                        "title": "Logo",
                        "description": "Logo file path of the team",
                        "type": "string",
                        "example": "/unity233625.jpg"
                    },
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Team"
                }
            },
...
            "TeamResource": {
                "title": "TeamResource",
                "description": "Team resource",
                "properties": {
                    "data": {
                        "title": "team Data",
                        "description": "Team Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Team"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TeamResource"
                }
            },
...
            "StoreTeamRequest": {
                "title": "Store Team request",
                "description": "Store Team request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the Team",
                        "type": "string",
                        "example": "unity"
                    },
                    "logo": {
                        "title": "Logo",
                        "description": "Logo file of the team",
                        "example": "/unity233625.jpg"
                    }
                },
                "type": "object"
            },
            "UpdatePlayerRequest": {
                "title": "Update player request",
                "description": "Update player request body data",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name of the player",
                        "type": "string",
                        "example": "Lionel"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name of the player",
                        "type": "string",
                        "example": "Messi"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image file of the player",
                        "example": "/playre123625.jpg"
                    }
                },
                "type": "object"
            },
            "UpdateTeamRequest": {
                "title": "Update Team request",
                "description": "Update Team request body data",
                "required": [],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the Team",
                        "type": "string",
                        "example": "unity"
                    },
                    "logo": {
                        "title": "Logo",
                        "description": "Logo file of the team",
                        "example": "/unity233625.jpg"
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        []
    ]
}